[a9d9e1c] [UPnP] split off UPnPServer to separate file. move some static class methods to
    helpers via internal header
